requirements.txt
Flask
flask-cors
Pillow
requests
python-dotenv
replicate
cloudinary
gunicorn
psycopg2-binary
opencv-python-headless
numpy
qrcode[pil]
